fix(url-builder): url builder is replaced for most security
