+++ .pre-commit-config.yaml
@@ -10,10 +10,10 @@ repos:
     rev: v0.5.0
     hooks:
       - id: go-fmt
-      - id: go-imports
+      # - id: go-imports
       - id: no-go-testing
       - id: golangci-lint
-      - id: go-unit-tests
+      # - id: go-unit-tests
       # - id: validate-toml
   - repo: https://github.com/alessandrojcm/commitlint-pre-commit-hook
     rev: v8.0.0
